added - font to project - navigation component - get location permission and WIP api calls

+++ app/src/main/java/ir/amirdaryabak/openweathermap/feature_home/data/remote/dto/geographic/CoordinatesDto.kt
@@ -3,8 +3,8 @@ package ir.amirdaryabak.openweathermap.feature_home.data.remote.dto.geographic
 import ir.amirdaryabak.openweathermap.feature_home.domain.entity.geographic.CoordinatesEntity
 
 data class CoordinatesDto(
-    val lat: Int,
-    val lon: Int
+    val lat: Double,
+    val lon: Double
 )
 
 fun CoordinatesDto.toCoordinatesEntity(): CoordinatesEntity {
